requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
psycopg2-binary==2.9.9
pandas==2.1.4
python-multipart==0.0.6
